Refactor BlogNavbar and Navbar components to utilize UserProfile component for user display and actions
